Further FMM Methods
Added neighbour finder and domain decomposition methods. Neighbour finder works, but is pretty ugly, domain decomposition creates a uniform grid at a given level, in future it will be possible for it to auto choose the level of subdivisions.

Also added a pair of new array extensions, ToUint to cast an array to uint, and Quotient to perform element wise division.